+++ billing/src/main/java/com/akademiaplus/membership/interfaceadapters/PaymentAdultStudentRepository.java
@@ -10,5 +10,5 @@ package com.akademiaplus.membership.interfaceadapters;
 import com.akademiaplus.billing.customerpayment.PaymentAdultStudentDataModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentAdultStudentRepository extends JpaRepository<PaymentAdultStudentDataModel, Integer> {
+public interface PaymentAdultStudentRepository extends JpaRepository<PaymentAdultStudentDataModel, Long> {
 }
