+++ billing/src/main/java/com/akademiaplus/membership/interfaceadapters/PaymentTutorRepository.java
@@ -10,5 +10,5 @@ package com.akademiaplus.membership.interfaceadapters;
 import com.akademiaplus.billing.customerpayment.PaymentTutorDataModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentTutorRepository extends JpaRepository<PaymentTutorDataModel, Integer> {
+public interface PaymentTutorRepository extends JpaRepository<PaymentTutorDataModel, Long> {
 }
