+++ course-management/src/main/java/com/akademiaplus/program/interfaceadapters/ScheduleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ScheduleRepository extends JpaRepository<ScheduleDataModel,  Integer> {
+public interface ScheduleRepository extends JpaRepository<ScheduleDataModel,  Long> {
 }
