+++ user-management/src/main/java/com/akademiaplus/collaborator/interfaceadapters/CollaboratorRepository.java
@@ -12,5 +12,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CollaboratorRepository extends JpaRepository<CollaboratorDataModel, Integer> {
+public interface CollaboratorRepository extends JpaRepository<CollaboratorDataModel, Long> {
 }
